#5 Rename module (methods → quantity)
